notification: make icon visible
The previous notification icon was a black exclamation point, which at
least on modern versions of Android is invisible.  Following the
iconography guidelines [1] and using the android-ui-utils notification
icon generator [2] I've made a simple clock icon that in addition to
being visible should also fit in reasonably well with other icons.

I added the Inkscape svg I used to create the icon to a new misc/ folder
because I wasn't sure where else would be good.

 [1] http://developer.android.com/design/style/iconography.html
 [2] http://android-ui-utils.googlecode.com/hg/asset-studio/dist/icons-notification.html