Fix bug in handling of multiple datapoints for a given day on the server.
For TT there should be only one datapoint per day on the beeminder server.
The way things were being handled, duplicates were being deleted from the
data fetched from the server, but the resulting empty entry in the array
was sometimes causing a (different) duplicate to be re-created later in the
script. This bug would only normally show up under circumstances where two
copies of TagTime were writing to the same graph, for example when it is being
run on two machines simultaneously.

This commit changes the deletion method to result in no duplicate entries
being created.